cmd: Add top subcommand.
This subcommand groups all *top gadgets like filetop or tcptop.
