+++ docs/guides/top/block-io.md
@@ -1,9 +1,9 @@
 ---
-title: 'The "biotop" gadget'
+title: 'The "top block-io" gadget'
 weight: 10
 ---
 
-The `biotop` gadget is used to trace block devices I/O.
+The `top block-io` gadget is used to trace block devices I/O.
 
 ## How to use it?
 
@@ -16,7 +16,7 @@ $ kubectl run test-pod --image busybox:latest sleep inf
 You can now use the gadget, but output will be empty:
 
 ```bash
-$ kubectl gadget biotop
+$ kubectl gadget top block-io
 NODE             NAMESPACE        POD              CONTAINER        PID     COMM             R/W MAJOR  MINOR  BYTES   TIME(µs) IOs
 ```
 
@@ -42,7 +42,7 @@ You can customize the information printed using `-o custom-columns=column0,...,c
 This command will only show the PID and command of the process which sent a signal:
 
 ```bash
-$ kubectl gadget biotop -o custom-columns=pid,comm
+$ kubectl gadget top block-io -o custom-columns=pid,comm
 PID     COMM
 7767    dd
 ```
@@ -50,7 +50,7 @@ PID     COMM
 The following command is the same as default printing:
 
 ```bash
-$ kubectl gadget biotop -o custom-columns=node,namespace,pod,container,pid,comm,r/w,major,minor,bytes,time,ios
+$ kubectl gadget top block-io -o custom-columns=node,namespace,pod,container,pid,comm,r/w,major,minor,bytes,time,ios
 NODE             NAMESPACE        POD              CONTAINER        PID     COMM             R/W MAJOR  MINOR  BYTES   TIME(µs) IOs
 minikube         default          test-pod         test-pod         7767    dd               W   0      0      1564672 3046     4
 ```
@@ -60,12 +60,12 @@ minikube         default          test-pod         test-pod         7767    dd  
 This gadget supports JSON output, for this simply use `-o json`:
 
 ```bash
-$ kubectl gadget biotop -o json
+$ kubectl gadget top block-io -o json
 []
 [{"node":"minikube","namespace":"default","pod":"test-pod","container":"test-pod","write":true,"bytes":2625536,"us":7075,"io":7,"mountnsid":4026532579,"pid":10310,"comm":"dd"}]
 []
 # You can use jq to make the output easier to read:
-$ kubectl gadget biotop -o json | jq
+$ kubectl gadget top block-io -o json | jq
 []
 [
   {
